Version 2.0
Hello. In this commit I implemented treeview and SQL database. Currently, the application works but test it out, see if you can find bugs, and experiment with it. If you can make it somewhat resistant to resizing since the prof is going to resize the app a bit and see what happens. Please use the version under v2 folder.